add enterprise logout link
