+++ src/config/enterprise.ts
@@ -1,0 +1,3 @@
+export default {
+  logoutPath: "/oidc/logout",
+};
